+++ R/mpcmp-package.R
@@ -58,14 +58,14 @@ NULL
 #' \item{finrest}{indicator variable for proposed change in ownership structure}
 #' \item{insthold}{percentage of stock held by institutions}
 #' \item{leglrest}{indicator variable for legal defence by lawsuit}
-#' \item{numbids}{number of bids recevied after the initial bid}
+#' \item{numbids}{number of bids received after the initial bid}
 #' \item{obs}{Identifier}
 #' \item{rearest}{indicator variable for proposed changes in asset structure}
 #' \item{regulatn}{indicator variable for Department of Justice intervention}
 #' \item{size}{total book value of assets in billions of dollars}
 #' \item{takeover}{Indicator. 1 if the company was being taken over}
 #' \item{weeks}{time in weeks between the initial and final offers}
-#' \item{whtknght}{indicator varible for management invitation 
+#' \item{whtknght}{indicator variable for management invitation 
 #' for friendly third-party bid}
 #' \item{sizesq}{book value squared}
 #' }
@@ -115,7 +115,7 @@ NULL
 #' \emph{Journal of Applied Statistics} \bold{41}, 2616--26.
 #' 
 #' @source
-#' Supplementary Content of Zevini et al. (2014): 
+#' Supplementary Content of Zeviani et al. (2014): 
 #' \url{http://www.leg.ufpr.br/doku.php/publications:papercompanions:zeviani-jas2014}
 #' 
 #' @examples 
@@ -150,6 +150,38 @@ NULL
 #' @examples 
 #' ## For examples see example(glm.cmp)
 NULL
+
+#' Sitophilus data set
+#'
+#' Ribeiro et al. (2013) carried out an experiment to assess the bioactivity of extracts from different parts (seeds, leaves and branches) of Annona mucosa (Annonaceae) to control Sitophilus zeamais (Coleoptera: Curculionidae), a major pest of stored maize/corn in Brazil.
+#'
+#' 10g of corn and 20 animals adults were placed in each Petri dish. Extracts prepared with different parts of mucosa or just water (control) were completely randomized with 10 replicates.
+#' 
+#' The numbers of emerged insects (progeny) after 60 days and their corresponding treatments were recorded in this dataset.
+#' 
+#' This dataset was obtained from the `cmpreg``package of Ribeiro Jr, 
+#' Zeviani & Demétrio (2019), which is based on the work of Ribeiro Junior et al. (2019).
+#' 
+#' This data set is also used to illustrate the syntax for regression on the dispersion. 
+#' @name sitophilus
+#' @format A data frame with 40 observations on 2 variables.
+#' \describe{
+#' \item{extract}{the treatment used}
+#' \item{ninsect}{number emerged insects (progeny)}
+#' }
+#' @docType data
+#' @keywords datasets
+#' @usage 
+#' data(sitophilus)
+#' @references 
+#' Ribeiro Junior, E.E., Zeviani, W.M., Bonat, W.H., Demétrio, C.G., & Hinde, J. (2019). Reparametrization of COM–Poisson regression models with applications in the analysis of experimental data. Statistical Modelling. https://doi.org/10.1177%2F1471082X19838651. 
+#' 
+#' Ribeiro, L.P., Vendramim, J.D., Bicalho, K.U., Andrade, M.S.,  Fernandes, J.B., Moral, R.A., & Demétrio C.G.B. (2013). Annona mucosa Jacq. (Annonaceae): A promising source of bioactive compounds against Sitophilus zeamais Mots. (Coleoptera: Curculionidae). \emph{Journal of Stored Products Research}, \bold{55}, 6-14.
+#' 
+#' @examples 
+#' ## For examples see example(glm.cmp)
+NULL
+
 
 ## usethis namespace: start
 #' @useDynLib mpcmp, .registration = TRUE